[Feature #160751808] Admin can post new menu
