--- server/dummyModels/foodItem.js
@@ -1,25 +1,0 @@
-const fastFoodItem = [
-  {
-    id: 1,
-    name: 'Extra large pizza',
-    description: 'Get your meals on the go in the most efficient manner. Our snacks are very tasty',
-    price: 1500,
-    imageurl: 'http://imagesavedhere1'
-  },
-  {
-    id: 2,
-    name: 'Burger and soda',
-    description: 'Get your meals on the go in the most efficient manner. Our snacks are very tasty',
-    price: 1500,
-    imageurl: 'http://imagesavedhere2'
-  },
-  {
-    id: 3,
-    name: 'Fried Chicken',
-    description: 'Get your meals on the go in the most efficient manner. Our snacks are very tasty',
-    price: 1000,
-    imageurl: 'http://imagesavedhere3'
-  }
-];
-
-export default fastFoodItem;
