--- server/dummyModels/orders.js
@@ -1,26 +1,0 @@
-const orders = [
-  {
-    id: 1,
-    userid: 3,
-    mealid: 6,
-    quantity: 3,
-    status: 1// accepted order
-  },
-  {
-    id: 2,
-    userid: 4,
-    mealid: 7,
-    quantity: 1,
-    status: 0// pending order
-  },
-  {
-    id: 3,
-    userid: 5,
-    mealid: 8,
-    quantity: 4,
-    status: 2// declined order
-  }
-];
-
-export default orders;
-
