+++ server/routes/index.js
@@ -1,11 +1,9 @@
 import users from './users';
 import menu from './menu';
-import myOrder from './orders';
 
 const allRoutes = {
   users,
-  menu,
-  myOrder
+  menu
 };
