--- server/tests/menu/postFoodItem.test.js
@@ -1,125 +1,0 @@
-import chai from 'chai';
-import chaiHttp from 'chai-http';
-import app from '../../app';
-
-chai.use(chaiHttp);
-const { expect } = chai;
-
-describe('POST /', () => {
-  let name;
-  let description;
-  let price;
-  let imageurl;
-
-  const exec = async () => {
-    try {
-      return await chai.request(app)
-        .post('/api/v1/foodItem')
-        .send({
-          name, description, price, imageurl
-        });
-    } catch (err) { throw err.message; }
-  };
-
-  beforeEach(() => {
-    name = 'myItem1';
-    description = 'a description message here';
-    price = '700';
-    imageurl = 'http://sometesturl';
-  });
-
-  it('should return a success status 201', async () => {
-    try {
-      const res = await exec();
-      expect(res.status).to.equal(201);
-      expect(res.body).to.be.an('object');
-      expect(res.body).to.have.property('message');
-    } catch (err) {
-      throw err.message;
-    }
-  });
-
-  it('should return a failure status for wrong food item name length 400', async () => {
-    try {
-      name = 'a';
-
-      const res = await exec();
-      expect(res.status).to.equal(400);
-      expect(res.body).to.be.an('object');
-      expect(res.body).to.have.property('message');
-    } catch (err) {
-      throw err.message;
-    }
-  });
-
-
-  it('should return a failure status for whitespace character', async () => {
-    try {
-      name = '      ';
-
-      const res = await exec();
-      expect(res.status).to.equal(400);
-      expect(res.body).to.be.an('object');
-      expect(res.body).to.have.property('message');
-      const errorMessage = 'One or more fields contained only whitespace';
-      expect(res.body).to.have.property('message', errorMessage);
-    } catch (err) {
-      throw err.message;
-    }
-  });
-
-  it('should return a failure status for incomplete query 400', async () => {
-    try {
-      name = '';
-
-      const res = await exec();
-      expect(res.status).to.equal(400);
-      expect(res.body).to.be.an('object');
-      expect(res.body).to.have.property('message');
-    } catch (err) {
-      throw err.message;
-    }
-  });
-
-  it('should return a failure status for not a number 400', async () => {
-    try {
-      price = 'addd';
-
-      const res = await exec();
-      expect(res.status).to.equal(400);
-      expect(res.body).to.be.an('object');
-      expect(res.body).to.have.property('message');
-    } catch (err) {
-      throw err.message;
-    }
-  });
-
-  it('should return a failure status for empty request', async () => {
-    try {
-      chai.request(app)
-        .post('/api/v1/foodItem')
-        .send({
-        })
-        .end((err, res) => {
-          expect(res.status).to.equal(400);
-          expect(res.body).to.be.an('object');
-          expect(res.body).to.have.property('message');
-        });
-    } catch (err) {
-      throw err.message;
-    }
-  });
-
-  it('should return a failure status for duplicate food item name 409', async () => {
-    try {
-      // duplicate of status 200
-      const res = await exec();
-      expect(res.status).to.equal(409);
-      expect(res.body).to.be.an('object');
-      expect(res.body).to.have.property('message');
-    } catch (err) {
-      throw err.message;
-    }
-  });
-});
-
